fix: do not consider the overview cache multiple times in a releasetrain (#1980)
This is a performance improvement that solves 2 issues with the overview
cache:
1) It reduces the amount of times we write the overview in a release
train from `n=number of apps` to `1 per env`. The 1 write contains
multiple changed apps.
2) When creating a new app or when undeploying, it does not force a
recalculation of the whole overview, but instead only updates the app.

Further cases that also force a complete recalculation of the overview
cache (like creating a new environment) are not handled in this PR.

Note that some functions like `CalculateWarnings` had to be moved to
another module but were not changed.

Ref: SRX-4WQBET
