+++ services/manifest-repo-export-service/pkg/service/version_test.go
@@ -254,7 +254,7 @@ func TestVersion(t *testing.T) {
 					App:     "test",
 					Env:     "development",
 					Version: &version,
-				}, 0)
+				}, 0, false)
 				err = repo.Apply(ctx, transaction, tc.Setup...)
 				if err != nil {
 					return err
